requirements.txt
# Gmail API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# AI and NLP
openai==1.3.6
langchain==0.0.340
transformers==4.35.2
torch==2.1.1
numpy==1.25.2

# Calendar Integration
google-calendar==0.7.2

# Email Processing
email-validator==2.1.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23

# Data Processing
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
